Removed unnecessary cairus.rs and implemented to_int() for Rgba
